move the tests in ters_dimensionless_units.py to test_units.py and add a test for checking missing mapping in the dimensionless mapping (not yet working as the checker has not yet been implemented)
